requirements.txt
corner==2.2.1
ipython==7.28.0
matplotlib==3.4.3
numexpr==2.7.3
numpy==1.21.2
opencv-python==4.5.1.48
pykdtree==1.3.4
pymc3==3.11.4
pyvista==0.32.1
seaborn==0.11.2
Theano-PyMC==1.1.2
tqdm==4.62.3
